[TASK] Make ExtensionConfig more tolerant to missing config
In case of a bug of TYPO3 where extension:setup does not
populate the EXTCONF array with initial (empty) extension
config, the ExtensionConfig service tries to access array
keys that do not exist.

While this is foremost a bug in TYPO3 extension:setup
handling, this workaround allows consumers of the
extension to still access their backend. Without this
patch, the backend will be inaccessible due to a
PHP "undefined array keys access" message.